Remove jetpack for package installation/management
In trying to deploy to Google Cloud, it couldn't find any packages.
Trial and error showed that not installing with jetpack fixed the
issue. Given that all this is dockerized and should work the same
everywhere, this makes no sense, but I can't be bothered trying
to figure out the root cause.