Seed a test DB for end-to-end tests
For browser-based testing that isn't useless, we'll need a DB
with test data to populate the data viz charts, which are the
main things that we need to test. For simplicity and speed, we're
using a static test DB, but we may want to populate it for each
test run in the future if outdated data proves to be an issue.