Refactor tip submission into its own class
Unfortunately, I can't find a way to write useful tests for this.
Either the test is as close to the real thing as possible,
which means it won't work for half the year when there's no tip
page; or I mock the browser object into ablivion, making the test
useless.