Add back docker container/image cleanup commands
I finally figured out the cause of the locale errors and a solution.
By default, ssh tries to transmit LC_* env vars for locale stuff.
The usual solution is modify ssh config on client or server side,
but I don't have that luxury, so I pass a blank ssh config file
and that seems to work fine.