Make is_correct null for unplayed matches
I try to avoid null when possible, because consistent data types
can reduce bugginess, but that impulse was misguided in this case,
because null is a meaningful value for prediction correctness: we
don't know whether it's correct yet, because the event hasn't
happened. Once a match has been played, then we can set is_correct
to true or false.