Fix tipping related bugs
There's no reason for data_import to depend on an env var when
we can get whether we're in production from Django settings,
which reduces the number of required env vars by one.

Also, I got a weird error from splinter while trying to run the
tip command, but fixing the version to the penultimate release
seems to have fixed it.