Fix predicted_margin cleaning & change to float
I've been meaning to make this value a float for awhile, because
the only reason I made it an integer was due to the fact that
footytips only accepts integer inputs for margins, but now we
submit to multiple competitions, so better to save the real
predicted margin and round it for submission only.

The impetus for this change was debugging how predicted margins
are calculated and cleaned. It turns out that full_clean calls
clean_fields first then clean, which resulted in coercing margins
to ints rather than rounding them.