Remove use of MLModelFactory from most tests
Creating & destroying MLModel records messes up use of
FullMatchFactory, because we iterate over ml_models to get
realistic associations for creating predictions. However, depending
on how objects are loaded we can get an outdated query to MLModel,
resulting in trying to fetch a record that no longer exists.

I left MLModelFactory in test_schema, because it was a huge pain
to try to remove it without breaking a bunch of other stuff due
to the assumption that all models have predictions for all matches,
which I'll probably have to fix eventually as I retire models.