Add gunicorn arguments to permit tips submissions
Doing all the tips submissions as part of the main process was
causing closed-connection errors, because gunicorn was timing out.
Increasing the timeout limit is a quick fix for this. I was also
having some issues seeing the logs, so added those args, and I
added a few workers as recommended by the documentation.