Refactor backfilling scores into separate command
This uses a separate data source from updating data for future
matches, and we'll eventually want to run it on a different
schedule. We only need to update fixtures once at the beginning
of the week, but we'll want to update scores throughout the round
to be able to display whether predictions were correct on the
website.