Use api functions in AWS Lambda handler
There are a lot of similar functions calling each other from
handler to api to data_import, so we'll probably want to
consolidate them eventually, but it's best to keep a separate,
controller-like module to make changing the web server
implementation as painless as possible.

As part of this, I made some changes to api and data_import to
remove unncessary param transformations and make the return values
consistent.