Replace calls to 'data' app with HTTP calls
This required adding a data importer to seed_db, since that's the
only case in which we allow the main app to trigger data transfers.
I replaced the data importing called in the backfill method with
a POST from 'tipping' to make sure all data updates from the
tipping process are controlled by the tipping service.