Update factories for more-realistic rounds & dates
In the past, I was able to get away with unrealistic randomness
with rounds and dates not being sequential and not even matching
up (i.e. later rounds having earlier dates and vice versa).
There are validation checks in some of the new functionality
that make tests flaky, so I added sequencing to make round numbers
and dates random within more realistic bounds.