Fix backfill when match is played but results aren't available
There's a bit of a delay between a match being played and scores
being updated on our match data source. We allow for this in
a validation, but didn't do an early return to skip later
validations and logic, which will raise an error one way or
another. This adds tests for some conditions that weren't being
tested, and fixes the validation to return if the scores aren't
available yet.