Fix cron settings for scheduled functions
Turns out that AWS cron starts Sunday at 1 and is always UTC
regardless of where the server is, so I updated the settings to
run on the same schedule as before with fixes to the cron values.
The one change in functionality is running prediction updates
through the weekend, because might as well.