Make datetimes cycle for shorter periods when late in the year
As we're getting later in the year, having six-month cycles for
match dates is running into the same issue as before, with our
reaching the end of the year and piling up matches on 31 Dec.
So, we shorten the cycle such that it's never longer than the
number of weeks that remain in the current year.