Add filtering query to Match model
FaunaDB is a bit limited in what queries it supports out of the
box, and I want to hold off on adding custom functions for
resolving queries. So I'm settling on basic value-matching filters
to do most of the work, then adding extra filtering on the
collection classes as needed, especially for more-complicated
filters like greater-/less-than.