Move all files other than index.html to 'root' folder
With a major version update, react-scripts changed the structure
of their build directory, which broke the previous 'mv' command
with empty file globs. Moving all except 'index.html' is more
robust to these sorts of changes, but requires a slightly
convoluted command in order to work in Docker's bash shell.