Allow for partial weeks for last week of the year
In the last week of the year, rounding the number of weeks left
results in a count of 0, which breaks the modulo math we use
later. This allows for a week count of 0, then we make the end
date just the end of the year. This will probably break on Dec 31
itself, but who cares.