Make sure score isn't 0 for played team matches
Scores for unplayed matches default to 0 (which was a mistake,
but I'm not going to fix it just now). This means that when we
check if a match has been updated with results since being played
we check for a non-zero score to make sure it was, so a score of
zero meant the match wasn't considered 'updated', so couldn't be
a draw.