Add skeleton for SQLAlchemy Dialect for Fauna
This creates the basic classes and methods needed to use a Fauna
as an SQLAlchemy Dialect, which will allow us to offload a lot of
model and DB query logic to SQLAlchemy rather than creating some
half-baked, homebrewed ORM ourselves.

Most of the boilerplate is copy-pasted from other 3rd-party SQLA
Dialects, whose similarities suggest that a common template exists
somewhere, but I didn't find one. So far, we can just create a
connection to a Fauna instance, but cannot execute any queries.