Allow FaunaDialect to create tables in the DB
This includes some fixes to the DB connection and integration
with SQLAlchemy that weren't caught by the basic connection test.
The main addition is the SQL parsing added to FaunadbClient.
It turns out that `sqlparse` has trouble with even some basic
SQL statements (e.g. incorrect splitting of tokens for column
definitions), so that might be an issue later, but I was able
to work around it for now.