Use default SQLAlchemy dialect exceptions
It's probably better to use the expected exceptions for
SQLAlchemy dialects. Also removed some of the assertions that
were more about working with the less-robust SQL traversal that
I was using at the start. To make sure we have the relevant info,
we always print the received SQL now before raising the original
error.