Add Match model and create migration
To get this to work, SQLAlchemy has to be able to fetch field
names, so I updated the INFORMATION_SCHEMA functionality. Also,
it turns out that SQLA prints the SQL query when you raise one
of it's native exceptions, so we don't have to print them in the
error message.