Do image caching in GitHub actions correctly
Turns out trying to leverage cached layers from a pulled Docker
image doesn't do anything, so I replaced that flow with a
combination of the Docker build actions and GitHub's native
caching action, which actually works. We're only caching the
Python apps, because those are the slowest to build, and I don't
want to run up against the memory limit for the cache.