Fix select queries when result is empty
When trying to select specific columns from a query without data,
SQLAlchemy was raising errors, because it couldn't find the
columns from teh query in the response. So, when we can't get
column names from a query response, we try to extract them from
the query itself.

Most of this is copied from FaunaClient's handling of SELECT
with specific columns/aliases. We'll get around to refactoring
it into some sort of SQL parser class eventually, but for now I
just want to get it working.