Install sqlalchemy-fauna as local package
It's not ready to be an independent package yet, but this will
allow us to extend functionality without needlessly coupling
the module with the tipping module.

I handled the http vs https for the DB calls with an env var,
but the SQLAlchemy-ElasticSearch package has separate dialects
for each. Both solutions are a bit clunky, but the env var one
is a bit simpler, so sticking with that for now.