Refactor insertion query to use one call only
With the separation of query translation from the execution
of those queries, it was a little awkward to have separate queries
for the collection metadata and creating the document.

We may still need a method to get a metadata query specifically,
because otherwise the unique constraint error message won't be
very useful.