Simplify select translation return & post-query response cleaning
I moved a lot of the logic for cleaning up responses to FQL,
which simplified the return value for translating SELECT queries
and made response handling across query types much more consistent.

This had some unanticipated knock-on effects. I had to make the
FQL response shaping extra complicated, because SQLAlchemy expects
key-value pairs in the same order as they appear in the query.
There is more room for refactoring and greater consistency across
query types, but this goes a long way toward removing special
cases in the client.