Consolidate queries that return Refs
With this change, all query types return the same value type.
Since CREATE queries require multiple FQL queries in a list,
we wrap the other query types in lists as well for consistency.
It's a bit clunky, but we'll probably want to support multiple
queries in a single SQL string eventually, so this will probably
be useful in the future.