Allow Match to create match objects from fixture data
I initially copied over the way I have this coded in the Django
ORM, but the active record style was kind of awkward, with
having to manage the SQLA session as an attribute of the Match
class. I didn't fully remove its dependance on the DB session,
but reduced it to data fetching, so we can handle transactions
and commits elsewhere.