Set default for score column to NULL
I've gone back and forth as to whether it's better to have
consistent data types or use null default values. In this case,
I think it's better to allow for nulls, since I want a clear
distinction between matches that haven't been played yet and those
that have. Also, conceptually, it makes more sense for a match's
scores to be null before it starts, and 0 once it's begun but
before anyone has scored.