Fix insertion & selection of documents with blank references
I was getting errors for null values where FQL doesn't permit
null values. This is partly due to my messy code, but partly to
how Select won't permit null values to be returned. I've settled
on a dummy NULL value and lots of 'if' statements as a work-around.

I also cleaned up the code in 'insert' a bit to make better use
of 'let' and lambdas, so it's a bit easier to follow what's
going on.