Make sure that upcoming match is on different day from previous
This test periodically doesn't raise the expected error, and I
think it's due to the next match being on the same day as the
first match, which would make the patched date the day before the
first match, which wouldn't raise, because the first match would
in fact be the next match. Since it's the first match of the fake
season, it would be saved in the DB without error.