Use subprocess instead of os to run shell commands
After writing this code, I've found out that subprocess is the
preferred method for running shell commands, as it's a wrapper
of the lower-level os methods that does more stuff (I guess?).
One of the benefits is raising an error on non-zero exit codes
so we can better handle problems. Given that integration tests
periodically fall over in CI for some DB-related reason, hopefully
this can help with debugging.