Create indices for all foreign reference combinations
Joining collections in Fauna is a bit tricky. I've settled on
creating more indices to be able to go from/to any combination
of foreign references to allow for more flexibility when chaining
joins from child to parent or vice versa. This means we don't
have to resort to paginate + map, which is more costly in terms
of speed an number of calls.