Add back foreign ref indices without target foreign ref values
Working a bit more on how to get chained joins to work in Fauna,
I've found that we just want to return refs often enough to make
having this type of index worthwhile. Otherwise, we have to keep
translating index matches to ref sets via joins that aren't
really necessary.