Save query column order in SQLQuery
With the introduction of support for multi-table queries, we need
to keep track of cross-table column order in the queries. By
default, it seems that SQLAlchemy selects columns by table, then
column, which works with the table-based column structure we had.
However, that's not a hard rule, and we need to guarantee that
response column order matches the query's column order to avoid
column label/value mismatches.