Hack together support for executemany
It was difficult to piece together what exactly SQLAlchemy
expects when calling executemany, but I was able to copy a
variation of the MySQL execution context to get the bare minimum
needed to make it work. This is needed, because SQLA really likes
grouping UPDATE calls together and having to call session.commit
every time you change any model's attributes seems unreasonable.