Implement updating of match results in api
The test code is getting a bit unruly, and I would rather have
api tests as the main integration tests while keeping the
different modules/classes as mostly unit tests, because test
setup for end-to-end api functions is a bit much.

Implementing this made me realise that the separation of
update_matches and update_match_results is mostly a holdover from
earlier implementation details, and there should just be one
version responsible for updating results.