Use information_schema collections for dialect queries
This updates existing methods for fetching DB metadata as well
as fully implementing some others that were previously empty.
All of these are based on examples I was able to find in the
SQLAlchemy source code, but I'm just winging it, so not sure if
the return values are exactly what's expected.