Simplify some conditionals
Was tempted to do some extra processing at the top of the query
building to normalize single-table and multi-table queries, so
we don't have to check for ref vs object later on. Though this
would simplify the code, it would result in much slower, costly
queries, because we would be fetching full documents for the
entire set, even if the query has a LIMIT clause or includes
aggregation functions, both of which cut down on how much data
we fetch.