Add position attribute to columns
Maintaining an ordered list of columns at the SQLQuery level has
led to some weird behaviour when trying to generate alias maps.
Since this was done to maintain the column order for the entire
query, rather than risk getting it mixed up nested in tables,
better to keep columns within tables and keep track of their
position with an attribute.