Use FilterGroup objects when bulding FQL queries to define document sets
Also updated model fixtures for the orderby + join test, because
the original implementation created a bug that applied the
order_by rule to every table being joined. This bug was only
raised when trying to order by a column that didn't exist on a
given table, which is why I added the 'game' attribute to the
Child model.