Fix bugs related to join tables & test properly
Integration in the tipping app revealed a bug from creating
join tables that resulted in duplicate indices, because we no
longer define foreign references per field, but rather per
collection.

In adding a join table to integration tests, I encountered
another bug, where we weren't supporting table aliases in queries.
Updating the Table, Column, and Filter to handle aliases was
enough. I'm not doing anything with the Table aliases yet, because
we don't need them in the results, but it might be something to
extend later.