Fix support or selecting tables other than one in FROM
Some holdover code that I had overlooked had the 'FROM' table
hard-coded as the key for the alias translation step when
processing SELECT results. This meant that selecting any tables
in a JOIN chain other than the first resulted in a blank response.

Changing this to be the table with selected columns fixes the
bug and allows selecting of any table in a JOIN chain. If we
eventually want to re-add support for selecting columns across
multiple tables, we'll probably have to base the key definition
on the given document ref.