Explicitly create join tables with model classes
The integration tests pass on local but fail on CI due to SQLA
generating a query that refers to two tables in the 'FROM' clause
('FROM foods, users_foods' instead of 'FROM foods JOIN users_foods').
I rebuilt the Docker image, and have the version of SQLA fixed,
so not sure where the discrepancy comes from, but easier to make
the test setup a bit hacky and add support for multiple 'FROM'
tables later.