Raise warnings in tests to catch them early
This is mostly for DeprecationWarning, but it's good for cleaning
up test logs and catching things before they become errors. I
also added the pytest config file, which isn't strictly necessary
since we're still running sqlalchemy-fauna as a module in tipping,
but it's good to have it in place so I don't forget later.