Add support for unions and intersections of FilterGroups
Now that I'm trying to implement support for nested WHERE clauses,
I need to rethink how filters are structured in the SQLQuery
object. Till now, FiltersGroups have been intersections of results,
with an implied union joining them together. Now, we need to be
able to handle unions and intersections of filter groups.

It's a little awkward for the time being with the nested
FilterGroups, but that's to hold off on updating the FQL functions
until I'm ready to add support for nested conditions, which will
require some big changes and introduction of recursion, so better
to do that in a separate PR.