Treat OR & AND filter logic consistently
This ended up having a lot of knock-on effects, because a lot of
logic implicitly depended on the convention of FilterGroups
representing intersections, and lists of FilterGroups representing
unions. There's a fair bit of duplication in the generation of
intersection and union FQL queries now, but I wanted to hold off
on refactoring till I got nested filtering logic working.