Add Comparison class
This allows us to move some logic from Filter to a separate class
that represents comparison logic. It started as a simple enum
and grew from there. More importantly, this allows us to use
instance methods while defining FQL queries rather than doing
so much low-level value checking & comparison.

For the 'checks_whether_*' methods, I just wrote them as normal
methods rather than trying to do a dynamic-definition thing,
because it's kind of janky in Python, and there aren't that many,
and it's not that big of a deal to add new ones as we add
support for new filter types.