Fill NaNs with 0 before returning cumulative metrics
This is a hacky workaround to account for the fact that we no
longer have a non-principal model and so don't have both
prediction types represented in the data. However, rewriting the
code to support NaN/None values is more trouble than it's worth
at the moment, and I just want to get the page loading again.