Install Chackra UI & dependencies & add boilerplate
Chakra UI requires a whole bunch of boilerplate to work with Remix,
so I copied all of that over from the documentation. One of the hooks
that Chakra UI uses has an empty dependency array, so I disabled
that rule in ESLint, because it's one I skip more often than adhere to.