Base schema on SQL dump & delete migrations
It ended up being much easier fixing up an SQL dump to work with
CockroachDB than write a seed script from scratch that reads a
JSON dump of the same data.

The schema file is much noiser as a result, but it's easy enough
to modify it and generate a migration that matches Prisma conventions.