Use Vercel Postgres instead of CockroachDB
I ran into a snag with Cockroach, because I already have an
organisation associated with my email, and they don't seem to
make it easy to have multiple projects/organisations associated
with a given account.

Rather than create another database under an unrelated project,
I'm going to try out Vercel's serverless RDB. It's a bit less
generous for the free tier, but still should be plenty for this
project.