Update remix-run dependencies to latest
The site is crashing due to a weird error inside remix-run,
so it's probably just too old, since I haven't updated packages
in about a year. Unfortunately, this required updating a bunch
of other packages and rewriting all the Remix boilerplate to match
the new version of the code generator.

I couldn't quite figure out how to incorporate Chakra UI into
entry.server.tsx, but it seems to work okay, so I commented out
the broken code and might come back to it later (but probably not).