Add Season table to DB
I want to be able to display historical predictions and model metrics,
but the current MlModel model doesn't have any concept of historical
states (e.g. the "principal" MLModel was always the current
principal model). Connecting models to the seasons in which they
generated predictions allows for us to show the predictions/metrics
for a model for a given season based on which ones were used in
competitions for that season.

I also added the season association to Match, since it makes sense
conceptually, and it will likely simplify season filters in SQL
queries down the road.