Fix unique constraint on MlModel
I had a unique constraint to restrict one principal model per year,
but it also had a filter to only apply to isPrincipal = TRUE,
which I don't have in this schema, so I just dropped it. I also
wanted to have a unique constraint for season/mlModel combos, so
I added that in.