Fetch round metrics one-at-a-time via select
This reduces the amount of data that we fetch on page load or
season change. I gave up on filtering metic names for the select
options, because it's a lot easier to just base it off the full
range of possible metrics rather than insert yet another action
in the more-complicated state management.